+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## v1.0.0
 _Work in progress._
 
+### Features
+- Add pagination support for the `list-projects` evcli command.
+
 ### Bug fixes
 - Remove initial blank line(s) in notification emails.
 - Remove invalid expiration date for `github/oauth2` identities.